refactor(receipt): split ReceiptDetail.tsx (1382 LOC) into container + parts [Phase3 C]
Pure mechanical extraction (data layer already on useQuery/useMutation).
src/components/ReceiptDetail.tsx → ReceiptDetail/index.tsx (container, 440
LOC: query + 3 mutations + state + handlers + derived fields + Metadata/md +
the inline re-extract button + render tree) and 12 presentational parts
under parts/ (SimpleBackBar, TopBar+MenuItem, AmountHero, StatusRow,
OriginalReceiptCollapsible+EmailHeaderStrip, NoteCard, LocationCard,
FieldsGrid+SmallFieldCard, LineItemsCard, ExtractionDetailsCollapsible,
ProcessingNote, banners). Single-consumer helpers travel with their
consumer; no shared.ts, no context, no state lifting. The load-bearing
iframe/secure-context comments stay verbatim in OriginalReceiptCollapsible.

git-renamed (history preserved); the sole importer resolves ./ReceiptDetail
→ the folder's index.tsx unchanged.

Behavior-identical. Verified in-browser (frontend-test-runner, 6/6): all
sections render, both collapsibles toggle, ⋯ menu + outside-click work, edit
mutation flows through the container (PATCH 200, reverted), both
original-receipt render paths (email iframe / photo img) work, no errors.
Build+tsc+lint(0).
