--- src/components/ReceiptDetail.tsx
@@ -1,1386 +1,0 @@
-import React, { useEffect, useRef, useState } from 'react';
-import { Link } from '@tanstack/react-router';
-import { brandLink, merchantLink } from '../lib/navLinks';
-import { MoreHorizontal, ChevronDown, ChevronRight } from 'lucide-react';
-import {
-  fetchReceiptDetail,
-  documentContentUrl,
-  documentRenderedUrl,
-  extractProblemMessage,
-  getTransaction,
-  postReExtractDocument,
-  toReceiptView,
-  voidTransaction,
-  restoreDocument,
-  type ReceiptView,
-  type BackendTransaction,
-  type BackendTransactionItem,
-  type ReExtractDocumentResult,
-} from '../lib/api';
-import { statusBadge } from '../lib/transactionStatus';
-import { cn } from '../lib/utils';
-import { MerchantIcon } from './MerchantIcon';
-import type { Category } from '../types';
-import EditReceiptModal from './EditReceiptModal';
-import ConfirmActionDialog from './ConfirmActionDialog';
-import DeleteReceiptDialog from './DeleteReceiptDialog';
-import DeletedBadge from './DeletedBadge';
-import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
-import { qk } from '../lib/queryKeys';
-import { removeTombstone } from '../lib/tombstones';
-
-interface ReceiptDetailProps {
-  receiptId: string;
-  onBack: () => void;
-  /** Bumped when a delete completes so the parent's transaction list
-   *  refetches. */
-  onAfterMutation?: () => void;
-}
-
-type Metadata = Record<string, unknown>;
-
-function md<T = unknown>(meta: Metadata | undefined, key: string): T | undefined {
-  if (!meta) return undefined;
-  const v = meta[key];
-  return v as T | undefined;
-}
-
-/**
- * Receipt detail — single-entry view in Variant B (Soft / Organic).
- * Follows docs/2026-05-10_Mockup_frontend_redesign-B-soft.html (fig.03).
- *
- * Functional surface is unchanged from the previous Material-3 version:
- *   - Auto-polls every 5s while status is draft/error.
- *   - Edit / Void / Delete / Restore actions kept (the mockup shows just
- *     Edit + Delete; Void and Restore are conditional flows that show up
- *     for posted/reconciled and tombstoned receipts respectively).
- *   - Renders line items, location map, raw OCR text, extraction quality
- *     when those metadata sub-objects exist.
- *
- * Data source: fetchReceiptDetail → real backend. No mocks, no fixtures.
- */
-export default function ReceiptDetail({ receiptId, onBack, onAfterMutation }: ReceiptDetailProps) {
-  const queryClient = useQueryClient();
-  const [activeDialog, setActiveDialog] = useState<'edit' | 'void' | 'delete' | null>(null);
-  const [restoreError, setRestoreError] = useState<string | null>(null);
-  // Re-extract state machine. `idle` armed; `pending` (~30-60s — the
-  // agent re-OCRs the image); `success` shows `changed_keys` toast;
-  // `error` flashes the problem-detail message. Mirrors the
-  // refresh-from-source pattern on MerchantDetail.
-  const [reExtractState, setReExtractState] = useState<
-    | { kind: 'idle' }
-    | { kind: 'pending' }
-    | { kind: 'success'; changedKeys: string[]; ocrChanged: boolean }
-    | { kind: 'error'; message: string }
-  >({ kind: 'idle' });
-
-  // The receipt detail (a Transaction + its documents) with its ETag embedded
-  // in the view. Auto-polls every 5s while the extractor is still working
-  // (status draft/error) via a conditional refetchInterval — replaces the old
-  // setInterval effect.
-  const {
-    data: receipt = null,
-    isLoading: loading,
-    error: queryError,
-  } = useQuery({
-    queryKey: qk.receipt(receiptId),
-    queryFn: () => fetchReceiptDetail(receiptId),
-    refetchInterval: (query) => {
-      const s = query.state.data?.status;
-      return s === 'draft' || s === 'error' ? 5000 : false;
-    },
-  });
-  const error = queryError ? extractProblemMessage(queryError) : null;
-
-  const invalidateReceipt = () =>
-    queryClient.invalidateQueries({ queryKey: qk.receipt(receiptId) });
-
-  // ETag write-back invariant: a PATCH/void response carries a FRESH ETag.
-  // Write the updated view straight into the cache so a subsequent edit sends
-  // the current If-Match. Invalidate-and-refetch would leave a stale-etag
-  // window in which a fast second edit 412s — so this is setQueryData, not
-  // invalidate. (EditReceiptModal calls this via onUpdated on PATCH success.)
-  const handleUpdated = (txn: BackendTransaction, etag: string | null) => {
-    queryClient.setQueryData(qk.receipt(receiptId), toReceiptView(txn, etag));
-  };
-
-  const voidMut = useMutation({
-    mutationFn: async (reason: string) => {
-      // Defensive: if the cached view lacks an etag, re-fetch a fresh one.
-      let etag = receipt?.etag ?? null;
-      if (!etag) {
-        const fresh = await getTransaction(receipt!.id);
-        if (!fresh.etag) throw new Error('No ETag — reload and retry.');
-        etag = fresh.etag;
-      }
-      return voidTransaction(receipt!.id, reason, etag);
-    },
-    onSuccess: () => {
-      setActiveDialog(null);
-      invalidateReceipt();
-      onAfterMutation?.();
-    },
-  });
-
-  const handleDeleted = () => {
-    setActiveDialog(null);
-    // A delete soft-tombstones the document; refresh the tombstone list too.
-    queryClient.invalidateQueries({ queryKey: qk.tombstones });
-    onAfterMutation?.();
-    onBack();
-  };
-
-  const restoreMut = useMutation({
-    mutationFn: () => restoreDocument(receipt!.documentId!),
-    onSuccess: () => {
-      removeTombstone(receipt!.documentId!);
-      invalidateReceipt();
-      queryClient.invalidateQueries({ queryKey: qk.tombstones });
-      onAfterMutation?.();
-    },
-    onError: (err: unknown) => setRestoreError(extractProblemMessage(err)),
-  });
-  const handleRestore = () => {
-    if (!receipt?.documentId) return;
-    setRestoreError(null);
-    restoreMut.mutate();
-  };
-
-  const reExtractMut = useMutation({
-    mutationFn: () => postReExtractDocument(receipt!.documentId!),
-    onMutate: () => setReExtractState({ kind: 'pending' }),
-    onSuccess: (result: ReExtractDocumentResult) => {
-      // Refresh the transaction so the UI reflects any field changes the
-      // agent committed (payee, occurred_on, occurred_at, etc).
-      invalidateReceipt();
-      onAfterMutation?.();
-      setReExtractState({
-        kind: 'success',
-        changedKeys: result.changed_keys,
-        ocrChanged: result.ocr_text_changed,
-      });
-      setTimeout(() => {
-        setReExtractState((s) => (s.kind === 'success' ? { kind: 'idle' } : s));
-      }, 6000);
-    },
-    onError: (err: unknown) =>
-      setReExtractState({ kind: 'error', message: extractProblemMessage(err) }),
-  });
-  const handleReExtract = () => {
-    if (!receipt?.documentId || reExtractMut.isPending) return;
-    reExtractMut.mutate();
-  };
-
-  if (loading) {
-    return (
-      <div className="space-y-4">
-        <SimpleBackBar onBack={onBack} />
-        <div className="py-16 text-center">
-          <p className="font-hand text-xl text-[var(--color-ink-muted)]">loading…</p>
-        </div>
-      </div>
-    );
-  }
-
-  if (error || !receipt) {
-    return (
-      <div className="space-y-4">
-        <SimpleBackBar onBack={onBack} />
-        <div className="rounded-[18px] border border-[var(--color-rule)] bg-[var(--color-surface)] py-12 text-center text-[var(--color-stamp)]">
-          {error || 'Receipt not found'}
-        </div>
-      </div>
-    );
-  }
-
-  // Pull extractor-stashed fields out of metadata.
-  const meta = (receipt as unknown as { documents: Array<{ extraction_meta?: Metadata }> });
-  const extraction = meta.documents[0]?.extraction_meta ?? undefined;
-  const txMeta = (receipt.postings[0] as unknown as { metadata?: Metadata })?.metadata;
-  const legacy: Metadata = { ...(txMeta ?? {}), ...(extraction ?? {}) };
-
-  const isProcessing = receipt.status === 'draft';
-  const tax = md<number>(legacy, 'tax');
-  const tip = md<number>(legacy, 'tip');
-  const rawText = md<string>(legacy, 'raw_text');
-  // #81: prefer transaction_items table (relational, with item_class /
-  // unit_price_minor / line_total_minor); fall back to legacy
-  // metadata.items JSON for transactions ingested before the lift.
-  const legacyItems = md<Array<{ name: string; quantity?: number; unit_price?: number; total_price?: number }>>(legacy, 'items');
-  const items: BackendTransactionItem[] = receipt.items.length > 0
-    ? receipt.items
-    : (legacyItems ?? []).map((i, idx) => ({
-        line_no: idx + 1,
-        raw_name: i.name,
-        normalized_name: null,
-        quantity: i.quantity ?? 1,
-        unit: null,
-        unit_price_minor: i.unit_price != null ? Math.round(i.unit_price * 100) : null,
-        line_total_minor: i.total_price != null ? Math.round(i.total_price * 100) : 0,
-        currency: receipt.currency,
-        item_class: 'other',
-        confidence: 'low',
-        line_type: 'product',
-        product_id: null,
-        tax_minor: null,
-        tip_share_minor: null,
-        discount_share_minor: null,
-        effective_total_minor: i.total_price != null ? Math.round(i.total_price * 100) : 0,
-      } satisfies BackendTransactionItem));
-  const confidence = md<number>(
-    (legacy.quality as Metadata | undefined) ?? {},
-    'confidence_score',
-  );
-  const warnings = md<string[]>(
-    (legacy.quality as Metadata | undefined) ?? {},
-    'warnings',
-  );
-  const merchantLabel = receipt.payee ?? receipt.narration ?? 'Unknown';
-
-  const canDelete = receipt.status !== 'voided';
-  const canVoid = receipt.status === 'posted' || receipt.status === 'reconciled';
-  const canEdit = receipt.status !== 'voided';
-
-  const primaryDoc = receipt.documents.find((d) => d.id === receipt.documentId) ?? receipt.documents[0];
-  const docDeletedAt = (primaryDoc as { deleted_at?: string | null } | undefined)?.deleted_at ?? null;
-  const isTombstoned = docDeletedAt != null;
-
-  const badge = statusBadge(receipt.status);
-  const lowConfidence = confidence != null && confidence < 0.6;
-
-  return (
-    <div className="space-y-6">
-      <TopBar
-        onBack={onBack}
-        isTombstoned={isTombstoned}
-        deletedAt={docDeletedAt}
-        isProcessing={isProcessing}
-        canEdit={canEdit}
-        canVoid={canVoid}
-        canDelete={canDelete}
-        restoring={restoreMut.isPending}
-        onEdit={() => setActiveDialog('edit')}
-        onVoid={() => setActiveDialog('void')}
-        onDelete={() => setActiveDialog('delete')}
-        onRestore={handleRestore}
-      />
-
-      <AmountHero
-        amount={receipt.total}
-        currency={receipt.currency}
-        merchant={isProcessing ? 'Processing…' : merchantLabel}
-        merchantBrandId={receipt.merchantBrandId}
-        category={receipt.category as Category | null}
-        occurredOn={receipt.occurred_on}
-        isProcessing={isProcessing}
-        voided={receipt.status === 'voided'}
-        brandTo={
-          // Merchant name in the hero → BrandPage (brand-level rollup
-          // across all locations). The per-location detail is reachable
-          // from the LocationCard below. Rendered as a real <a href> so
-          // it opens in a new tab / split view on right-click.
-          receipt.merchantBrandId ? brandLink(receipt.merchantBrandId) : undefined
-        }
-      />
-
-      <StatusRow
-        badge={badge}
-        paymentMethod={receipt.paymentMethod ?? null}
-        source={primaryDoc?.kind ?? null}
-      />
-
-      {isProcessing && <ProcessingNote />}
-
-      {restoreError && (
-        <Banner tone="error">Restore failed: {restoreError}</Banner>
-      )}
-
-      <FieldsGrid
-        payment={receipt.paymentMethod ?? null}
-        tax={tax}
-        tip={tip}
-        isProcessing={isProcessing}
-      />
-
-      {!isProcessing && (
-        <LocationCard
-          place={receipt.place}
-          merchantId={receipt.merchantId}
-          payee={receipt.payee ?? null}
-        />
-      )}
-
-      {!isProcessing && items.length > 0 && (
-        <LineItemsCard items={items} currency={receipt.currency} />
-      )}
-
-      {receipt.narration && !isProcessing && (
-        <NoteCard text={receipt.narration} />
-      )}
-
-      {/* Related Email slot — populated once Gmail integration (#34) ships.
-       *  Hidden when there are no matches (no skeleton, no placeholder). */}
-
-      {!isProcessing && receipt.documentId && (
-        <OriginalReceiptCollapsible
-          documentId={receipt.documentId}
-          kind={primaryDoc?.kind ?? null}
-          sourceMeta={
-            (primaryDoc as { source_meta?: Record<string, unknown> | null } | undefined)
-              ?.source_meta ?? null
-          }
-        />
-      )}
-
-      {/* Re-extract affordance. Only on active (non-voided) receipts
-          that have a linked document. Wall-time is ~30-60s for vision
-          OCR, so we make the pending state visible. */}
-      {!isProcessing &&
-        receipt.documentId &&
-        receipt.status !== 'voided' && (
-          <div className="space-y-2">
-            <button
-              type="button"
-              onClick={handleReExtract}
-              disabled={reExtractState.kind === 'pending'}
-              className={cn(
-                'group inline-flex items-center gap-2 text-[11px] uppercase tracking-[0.16em]',
-                'text-[var(--color-ink-muted)] hover:text-[var(--color-terracotta)]',
-                'transition-colors disabled:opacity-50 disabled:cursor-wait',
-              )}
-              title="Re-run OCR with the current model and prompt"
-            >
-              <span className="font-display italic text-base leading-none text-[var(--color-terracotta)] group-hover:translate-x-px transition-transform">
-                ↺
-              </span>
-              {reExtractState.kind === 'pending'
-                ? 're-extracting… (~30-60s)'
-                : 'Re-extract'}
-            </button>
-
-            {reExtractState.kind === 'success' && (
-              <ReExtractBanner
-                tone="success"
-                onDismiss={() => setReExtractState({ kind: 'idle' })}
-              >
-                {reExtractState.changedKeys.length === 0 && !reExtractState.ocrChanged
-                  ? 'No changes — the agent produced the same output.'
-                  : reExtractState.changedKeys.length === 0
-                    ? 'OCR text refreshed; no transaction fields changed.'
-                    : `Updated ${reExtractState.changedKeys.join(', ')}.`}
-              </ReExtractBanner>
-            )}
-            {reExtractState.kind === 'error' && (
-              <ReExtractBanner
-                tone="error"
-                onDismiss={() => setReExtractState({ kind: 'idle' })}
-              >
-                {reExtractState.message}
-              </ReExtractBanner>
-            )}
-          </div>
-        )}
-
-      {!isProcessing && (rawText || confidence != null) && (
-        <ExtractionDetailsCollapsible
-          rawText={rawText}
-          confidence={confidence}
-          warnings={warnings}
-          defaultOpen={lowConfidence}
-        />
-      )}
-
-      {/* Dialogs */}
-      <EditReceiptModal
-        isOpen={activeDialog === 'edit'}
-        onClose={() => setActiveDialog(null)}
-        receipt={receipt}
-        onUpdated={handleUpdated}
-        onStale={invalidateReceipt}
-      />
-
-      <ConfirmActionDialog
-        isOpen={activeDialog === 'void'}
-        onClose={() => setActiveDialog(null)}
-        title="Void this receipt?"
-        message={
-          <>
-            <p>
-              Voiding creates a reversing entry in the ledger — the original transaction stays,
-              but its balance cancels out. Use this for posted receipts you can't simply delete.
-            </p>
-            <p className="mt-2 text-xs text-[var(--color-ink-muted)]">
-              This action can be reversed only by creating a new offsetting transaction.
-            </p>
-          </>
-        }
-        confirmLabel="Void receipt"
-        destructive
-        requireReason
-        reasonPlaceholder="Why are you voiding this? (optional)"
-        onConfirm={(reason) => voidMut.mutateAsync(reason).then(() => undefined)}
-      />
-
-      <DeleteReceiptDialog
-        isOpen={activeDialog === 'delete'}
-        onClose={() => setActiveDialog(null)}
-        documentId={receipt.documentId}
-        transactionId={receipt.id}
-        transactionEtag={receipt.etag}
-        isReconciled={receipt.status === 'reconciled'}
-        onDeleted={handleDeleted}
-      />
-    </div>
-  );
-}
-
-/* ── Pieces ───────────────────────────────────────────────────── */
-
-function SimpleBackBar({ onBack }: { onBack: () => void }) {
-  return (
-    <div className="flex items-center text-[11px] tracking-[0.16em] uppercase text-[var(--color-ink-muted)]">
-      <button
-        type="button"
-        onClick={onBack}
-        className="flex items-center gap-2 hover:text-[var(--color-ink)] transition-colors"
-      >
-        <span className="font-display italic text-lg leading-none text-[var(--color-terracotta)]">←</span>
-        Back
-      </button>
-    </div>
-  );
-}
-
-function TopBar({
-  onBack,
-  isTombstoned,
-  deletedAt,
-  isProcessing,
-  canEdit,
-  canVoid,
-  canDelete,
-  restoring,
-  onEdit,
-  onVoid,
-  onDelete,
-  onRestore,
-}: {
-  onBack: () => void;
-  isTombstoned: boolean;
-  deletedAt: string | null;
-  isProcessing: boolean;
-  canEdit: boolean;
-  canVoid: boolean;
-  canDelete: boolean;
-  restoring: boolean;
-  onEdit: () => void;
-  onVoid: () => void;
-  onDelete: () => void;
-  onRestore: () => void;
-}) {
-  const [menuOpen, setMenuOpen] = useState(false);
-  const menuRef = useRef<HTMLDivElement | null>(null);
-
-  useEffect(() => {
-    if (!menuOpen) return;
-    const onDoc = (e: MouseEvent) => {
-      if (menuRef.current && !menuRef.current.contains(e.target as Node)) setMenuOpen(false);
-    };
-    document.addEventListener('mousedown', onDoc);
-    return () => document.removeEventListener('mousedown', onDoc);
-  }, [menuOpen]);
-
-  return (
-    <div className="flex items-center justify-between text-[11px] tracking-[0.16em] uppercase text-[var(--color-ink-muted)]">
-      <button
-        type="button"
-        onClick={onBack}
-        className="flex items-center gap-2 hover:text-[var(--color-ink)] transition-colors"
-      >
-        <span className="font-display italic text-lg leading-none text-[var(--color-terracotta)]">←</span>
-        Back
-      </button>
-      <div className="flex items-center gap-3">
-        {isTombstoned && <DeletedBadge deletedAt={deletedAt} />}
-        {isTombstoned ? (
-          <button
-            type="button"
-            onClick={onRestore}
-            disabled={restoring}
-            className={cn(
-              'rounded-full px-3 py-1.5 text-[11px] font-medium tracking-[0.14em] uppercase',
-              'bg-[var(--color-terracotta)] text-white hover:bg-[var(--color-terracotta-deep)]',
-              'transition-colors disabled:opacity-50 disabled:cursor-not-allowed',
-            )}
-          >
-            {restoring ? 'Restoring…' : 'Restore'}
-          </button>
-        ) : (
-          !isProcessing && (
-            <div ref={menuRef} className="relative">
-              <button
-                type="button"
-                onClick={() => setMenuOpen((s) => !s)}
-                aria-label="Receipt actions"
-                className={cn(
-                  'flex h-8 w-8 items-center justify-center rounded-full',
-                  'border border-[var(--color-rule)] bg-[var(--color-surface)]',
-                  'hover:border-[var(--color-ink)]/30 transition-colors',
-                )}
-              >
-                <MoreHorizontal size={16} className="text-[var(--color-ink)]" />
-              </button>
-              {menuOpen && (
-                <div
-                  className={cn(
-                    'absolute right-0 z-30 mt-2 min-w-[180px] p-1',
-                    'rounded-[14px] bg-[var(--color-surface)] border border-[var(--color-rule)]',
-                    'shadow-[0_12px_32px_-10px_rgba(45,37,32,0.18)]',
-                  )}
-                >
-                  <MenuItem
-                    label="Edit fields"
-                    disabled={!canEdit}
-                    onClick={() => {
-                      setMenuOpen(false);
-                      onEdit();
-                    }}
-                  />
-                  {canVoid && (
-                    <MenuItem
-                      label="Void receipt"
-                      onClick={() => {
-                        setMenuOpen(false);
-                        onVoid();
-                      }}
-                    />
-                  )}
-                  <MenuItem
-                    label="Delete…"
-                    disabled={!canDelete}
-                    destructive
-                    onClick={() => {
-                      setMenuOpen(false);
-                      onDelete();
-                    }}
-                  />
-                </div>
-              )}
-            </div>
-          )
-        )}
-      </div>
-    </div>
-  );
-}
-
-function MenuItem({
-  label,
-  onClick,
-  disabled,
-  destructive,
-}: {
-  label: string;
-  onClick: () => void;
-  disabled?: boolean;
-  destructive?: boolean;
-}) {
-  return (
-    <button
-      type="button"
-      onClick={onClick}
-      disabled={disabled}
-      className={cn(
-        'w-full text-left px-3 py-2 rounded-[10px] text-sm transition-colors normal-case tracking-normal',
-        destructive
-          ? 'text-[var(--color-stamp)] hover:bg-[var(--color-stamp)]/8'
-          : 'text-[var(--color-ink)] hover:bg-[var(--color-paper-deep)]',
-        'disabled:opacity-40 disabled:cursor-not-allowed',
-      )}
-    >
-      {label}
-    </button>
-  );
-}
-
-function AmountHero({
-  amount,
-  currency,
-  merchant,
-  merchantBrandId,
-  category,
-  occurredOn,
-  isProcessing,
-  voided,
-  brandTo,
-}: {
-  amount: number;
-  currency: string;
-  merchant: string;
-  merchantBrandId: string | null;
-  category: Category | null;
-  occurredOn: string;
-  isProcessing: boolean;
-  voided: boolean;
-  /** Link target for the merchant name → BrandPage. Undefined renders a
-   *  plain <h1> (no link) — e.g. while processing or with no brand. */
-  brandTo?: ReturnType<typeof brandLink>;
-}) {
-  const merchantClass = 'font-display italic font-medium text-2xl sm:text-3xl leading-tight';
-  // FE#48: small square icon next to the merchant name. Skipped while
-  // processing (the row says "Processing…", not a real merchant yet).
-  const showIcon = !isProcessing;
-  return (
-    <div className="text-center pt-2">
-      <p
-        className={cn(
-          'font-display italic font-medium tracking-tight tnum',
-          'text-[3.25rem] sm:text-[4rem] leading-none',
-          voided && 'line-through text-[var(--color-ink-muted)]',
-        )}
-      >
-        {isProcessing ? '—' : `$${amount.toFixed(2)}`}
-      </p>
-      <p className="mt-1 text-[11px] tracking-[0.14em] uppercase text-[var(--color-ink-muted)]">
-        {currency}
-      </p>
-      {brandTo ? (
-        <Link
-          {...brandTo}
-          className={cn(
-            'mt-4 inline-flex items-center gap-2 transition-colors hover:text-[var(--color-terracotta)]',
-            merchantClass,
-          )}
-        >
-          {showIcon && (
-            <MerchantIcon brandId={merchantBrandId} category={category} size={28} />
-          )}
-          <span className="inline-flex items-baseline gap-1">
-            {merchant}
-            <span className="font-display italic text-base leading-none text-[var(--color-terracotta)]">→</span>
-          </span>
-        </Link>
-      ) : (
-        <h1 className={cn('mt-4 inline-flex items-center gap-2', merchantClass)}>
-          {showIcon && (
-            <MerchantIcon brandId={merchantBrandId} category={category} size={28} />
-          )}
-          <span>{merchant}</span>
-        </h1>
-      )}
-      <p className="mt-1 text-[13px] text-[var(--color-ink-muted)]">
-        {formatDateLong(occurredOn)}
-      </p>
-    </div>
-  );
-}
-
-function sourceLabel(
-  kind: string | null | undefined,
-): { glyph: string; label: string } | null {
-  if (kind === 'receipt_email') return { glyph: '✉︎', label: 'via email' };
-  if (kind === 'receipt_pdf' || kind === 'statement_pdf')
-    return { glyph: '⌗', label: 'via pdf' };
-  return null;
-}
-
-function StatusRow({
-  badge,
-  paymentMethod,
-  source,
-}: {
-  badge: ReturnType<typeof statusBadge>;
-  paymentMethod: string | null;
-  source?: string | null;
-}) {
-  const src = sourceLabel(source);
-  if (!badge && !paymentMethod && !src) return null;
-  return (
-    <div className="flex items-center justify-center gap-2 text-[12px] text-[var(--color-ink-muted)]">
-      {src && (
-        <span
-          className={cn(
-            'inline-flex items-center gap-1 px-2 py-0.5 rounded-full text-[11px] font-medium',
-            'bg-[var(--color-terracotta-soft)] text-[var(--color-terracotta-deep)]',
-          )}
-        >
-          <span aria-hidden="true">{src.glyph}</span>
-          {src.label}
-        </span>
-      )}
-      {src && (badge || paymentMethod) && <span aria-hidden="true">·</span>}
-      {badge && (
-        <span
-          className={cn(
-            'inline-flex items-center px-2 py-0.5 rounded-full text-[11px] font-medium',
-            badge.tone === 'red' && 'bg-[var(--color-stamp)]/10 text-[var(--color-stamp)]',
-            badge.tone === 'green' && 'bg-[color:rgba(52,168,83,0.12)] text-[color:rgb(52,168,83)]',
-            badge.tone === 'muted' && 'bg-[var(--color-paper-deep)] text-[var(--color-ink-muted)]',
-          )}
-        >
-          {badge.label}
-        </span>
-      )}
-      {badge && paymentMethod && <span aria-hidden="true">·</span>}
-      {paymentMethod && (
-        <span>
-          {paymentMethod.replace(/_/g, ' ').replace(/\b\w/g, (c) => c.toUpperCase())}
-        </span>
-      )}
-    </div>
-  );
-}
-
-function formatReceived(iso: string): string {
-  const d = new Date(iso);
-  if (Number.isNaN(d.getTime())) return iso;
-  return d.toLocaleDateString(undefined, {
-    year: 'numeric',
-    month: 'short',
-    day: 'numeric',
-  });
-}
-
-/** Email message-header strip shown above the rendered email body, so the
- *  fold reads as "the original email" rather than a raw HTML blob (#76). */
-function EmailHeaderStrip({
-  sender,
-  subject,
-  receivedAt,
-}: {
-  sender: string | null;
-  subject: string | null;
-  receivedAt: string | null;
-}) {
-  return (
-    <div className="rounded-[10px] border border-[var(--color-rule)] bg-[var(--color-surface)] px-4 py-3">
-      {subject && (
-        <p className="font-display italic font-medium text-[15px] leading-snug text-[var(--color-ink)]">
-          {subject}
-        </p>
-      )}
-      <div className="mt-1 flex flex-wrap items-center gap-x-2 gap-y-0.5 text-[12px] text-[var(--color-ink-muted)]">
-        {sender && <span className="min-w-0 truncate">{sender}</span>}
-        {sender && receivedAt && <span aria-hidden="true">·</span>}
-        {receivedAt && <span className="tnum">{formatReceived(receivedAt)}</span>}
-      </div>
-    </div>
-  );
-}
-
-function OriginalReceiptCollapsible({
-  documentId,
-  kind,
-  sourceMeta,
-}: {
-  documentId: string;
-  kind?: string | null;
-  sourceMeta?: Record<string, unknown> | null;
-}) {
-  const [open, setOpen] = useState(false);
-  const isEmail = kind === 'receipt_email';
-  const isPdf = kind === 'receipt_pdf' || kind === 'statement_pdf';
-  const sender = (sourceMeta?.sender as string | undefined) ?? null;
-  const subject = (sourceMeta?.subject as string | undefined) ?? null;
-  const receivedAt = (sourceMeta?.received_at as string | undefined) ?? null;
-  return (
-    <div className="rounded-[18px] border border-[var(--color-rule)] bg-[var(--color-surface)] overflow-hidden">
-      <button
-        type="button"
-        onClick={() => setOpen((s) => !s)}
-        className="w-full px-5 py-4 flex items-center justify-between text-left hover:bg-[var(--color-paper-deep)]/30 transition-colors"
-      >
-        <span className="font-display italic font-medium text-lg leading-none">
-          {isEmail ? 'Original email' : 'Original receipt'}
-        </span>
-        {open ? <ChevronDown size={18} className="text-[var(--color-ink-muted)]" /> : <ChevronRight size={18} className="text-[var(--color-ink-muted)]" />}
-      </button>
-      {open && (
-        <div
-          className="p-4"
-          style={{
-            background:
-              'linear-gradient(180deg, rgba(245, 230, 195, 0.4) 0%, rgba(201, 123, 92, 0.06) 100%), var(--color-surface)',
-          }}
-        >
-          {isEmail ? (
-            <div className="space-y-3">
-              {(sender || subject || receivedAt) && (
-                <EmailHeaderStrip
-                  sender={sender}
-                  subject={subject}
-                  receivedAt={receivedAt}
-                />
-              )}
-              {/* Faithful render of the sender's HTML, isolated in a
-                  sandboxed iframe (no scripts / no same-origin). The
-                  backend serves it with a strict CSP; this is the second
-                  containment layer. Fixed height + internal scroll since
-                  email heights vary wildly. */}
-              <iframe
-                src={documentRenderedUrl(documentId)}
-                title="Original email"
-                sandbox=""
-                referrerPolicy="no-referrer"
-                className="block w-full rounded-[10px] border border-[var(--color-rule)] bg-white"
-                style={{ height: 520 }}
-              />
-            </div>
-          ) : isPdf ? (
-            <div className="space-y-3">
-              {/* PDFs can't render in an <img> tag — embed the browser's
-                  built-in PDF viewer via an iframe pointing at the raw
-                  /content stream (served as application/pdf). No sandbox:
-                  this is our own same-origin file, and a strict sandbox
-                  blocks the viewer. The link below is the fallback for
-                  mobile browsers (iOS Safari) that won't render PDFs
-                  inline. */}
-              <iframe
-                src={documentContentUrl(documentId)}
-                title="Original PDF"
-                className="block w-full rounded-[10px] border border-[var(--color-rule)] bg-white"
-                style={{ height: 520 }}
-              />
-              <a
-                href={documentContentUrl(documentId)}
-                target="_blank"
-                rel="noopener noreferrer"
-                className="inline-block text-sm font-medium text-[var(--color-terracotta)] hover:underline"
-              >
-                Open PDF in new tab
-              </a>
-            </div>
-          ) : (
-            <img
-              src={documentContentUrl(documentId)}
-              alt="Receipt"
-              className="block max-w-full max-h-[500px] object-contain mx-auto rounded-[10px]"
-              onError={(e) => {
-                (e.target as HTMLImageElement).style.display = 'none';
-              }}
-            />
-          )}
-        </div>
-      )}
-    </div>
-  );
-}
-
-function NoteCard({ text }: { text: string }) {
-  return (
-    <div
-      className="relative rounded-[16px] px-4 py-4 leading-snug"
-      style={{ background: 'var(--color-butter)' }}
-    >
-      <span
-        className={cn(
-          'absolute -top-2 left-4 inline-block rounded-full px-2 py-[3px]',
-          'bg-[var(--color-terracotta)] text-white',
-          'text-[10px] font-medium tracking-[0.16em] uppercase',
-        )}
-      >
-        your note
-      </span>
-      <p className="font-hand text-lg text-[var(--color-ink)]">{text}</p>
-    </div>
-  );
-}
-
-// FE#15 + design #35 (pastel-pillow): persistent Location card. Whole card
-// click opens Google Maps for the actual merchant (uses google_place_id
-// + payee), not just the address. "View all visits" remains the in-app
-// path to MerchantDetail.
-function LocationCard({
-  place,
-  merchantId,
-  payee,
-}: {
-  place: ReceiptView['place'];
-  merchantId: string | null;
-  payee: string | null;
-}) {
-  const addr = place?.formatted_address?.trim() || null;
-  const cityState = (() => {
-    if (!addr) return null;
-    const parts = addr.split(',').map((s) => s.trim()).filter(Boolean);
-    if (parts.length < 2) return null;
-    if (place?.country_code === 'US' && parts.length >= 4) {
-      const city = parts[parts.length - 3];
-      const stateZip = parts[parts.length - 2];
-      const st = stateZip?.split(/\s+/)[0];
-      return st ? `${city}, ${st}` : city ?? null;
-    }
-    return parts[parts.length - 2] ?? null;
-  })();
-
-  // Build a Maps URL that resolves to the merchant's place card, not just
-  // the address. We only pass `query_place_id` when we believe the
-  // place_id points to a real business — when the backend's geocoder
-  // matched a premise / street / postal code instead, the place_id
-  // resolves to a bare address and Google faithfully shows that. In
-  // those cases we drop the place_id and rely on Google's text search,
-  // which finds the actual merchant at that address (e.g. "Tokyo
-  // Central 1740 Artesia Blvd, Gardena, CA").
-  const GEOCODING_FALLBACK_TYPES = new Set([
-    'premise',
-    'subpremise',
-    'street_address',
-    'route',
-    'intersection',
-    'postal_code',
-    'plus_code',
-    'locality',
-    'sublocality',
-    'neighborhood',
-    'administrative_area_level_1',
-    'administrative_area_level_2',
-    'administrative_area_level_3',
-    'country',
-  ]);
-  const trustPlaceId =
-    !!place?.google_place_id &&
-    !!place?.primary_type &&
-    !GEOCODING_FALLBACK_TYPES.has(place.primary_type);
-
-  const mapsUrl = (() => {
-    if (!place) return null;
-    const name = payee?.trim() || place.display_name_en || '';
-    const q = encodeURIComponent(`${name} ${addr ?? ''}`.trim());
-    if (!q) return null;
-    const base = `https://www.google.com/maps/search/?api=1&query=${q}`;
-    return trustPlaceId
-      ? `${base}&query_place_id=${encodeURIComponent(place.google_place_id)}`
-      : base;
-  })();
-  const openMaps = () => {
-    if (mapsUrl) window.open(mapsUrl, '_blank', 'noopener,noreferrer');
-  };
-
-  if (!place) {
-    return (
-      <div className="rounded-[14px] border border-[var(--color-rule)] bg-[var(--color-surface)] px-4 py-3">
-        <p className="text-[11px] font-medium tracking-[0.14em] uppercase text-[var(--color-ink-muted)]">
-          Location
-        </p>
-        <p className="mt-2 text-[13px] text-[var(--color-ink-muted)]">
-          Location unavailable —{' '}
-          <span className="italic">geocoding may be disabled or failed.</span>
-        </p>
-      </div>
-    );
-  }
-
-  return (
-    <div
-      role="button"
-      tabIndex={0}
-      onClick={openMaps}
-      onKeyDown={(e) => {
-        if (e.key === 'Enter' || e.key === ' ') {
-          e.preventDefault();
-          openMaps();
-        }
-      }}
-      className="relative cursor-pointer overflow-hidden rounded-[28px] px-5 py-5 transition-transform duration-150 ease-out hover:-translate-y-[1px] focus:outline-none focus-visible:ring-2 focus-visible:ring-[var(--color-terracotta)]"
-      style={{
-        background: 'linear-gradient(135deg,#FDE2E4 0%,#E2F0CB 50%,#CDE7F7 100%)',
-        boxShadow:
-          'inset 0 4px 0 rgba(255,255,255,0.6), 0 14px 30px -10px rgba(180,140,180,0.32), 0 1px 2px rgba(60,40,20,0.04)',
-        fontFamily: 'Quicksand, ui-sans-serif, system-ui, sans-serif',
-        color: '#3A3550',
-      }}
-    >
-      <span
-        aria-hidden="true"
-        className="pointer-events-none absolute -top-[70px] -right-[60px] h-[180px] w-[180px] rounded-full"
-        style={{ background: 'rgba(255,255,255,0.55)', filter: 'blur(14px)' }}
-      />
-      <span
-        aria-hidden="true"
-        className="pointer-events-none absolute -bottom-[30px] left-[30%] h-[100px] w-[100px] rounded-full"
-        style={{ background: 'rgba(255,180,200,0.35)', filter: 'blur(20px)' }}
-      />
-
-      <p
-        className="relative text-[11px] font-medium tracking-[0.18em] uppercase"
-        style={{
-          color: '#7A6F92',
-          fontFamily: 'Plus Jakarta Sans, ui-sans-serif, system-ui, sans-serif',
-        }}
-      >
-        Location
-      </p>
-
-      <div className="relative mt-3 flex items-center gap-4">
-        <div
-          aria-hidden="true"
-          className="shrink-0 inline-flex items-center justify-center"
-          style={{
-            width: 60,
-            height: 60,
-            fontSize: 28,
-            background: 'linear-gradient(135deg,#FFB5C2,#FFE2A8)',
-            borderRadius: '42% 58% 65% 35% / 38% 50% 50% 62%',
-            boxShadow:
-              '0 6px 16px -6px rgba(255,140,160,0.45), inset 0 2px 0 rgba(255,255,255,0.5)',
-            animation: 'pastel-blob 8s ease-in-out infinite',
-          }}
-        >
-          📍
-        </div>
-        <div className="min-w-0 flex-1">
-          {cityState && (
-            <p
-              className="text-[20px] font-bold leading-tight"
-              style={{ color: '#3A3550', letterSpacing: '-0.01em' }}
-            >
-              {cityState}
-            </p>
-          )}
-          {addr && (
-            <p
-              className="mt-0.5 text-[13px] font-medium leading-snug"
-              style={{ color: '#7A7090' }}
-            >
-              {addr}
-            </p>
-          )}
-          {!cityState && !addr && (
-            <p className="text-[13px] font-medium" style={{ color: '#7A7090' }}>
-              Geocoded, but no street address returned.
-            </p>
-          )}
-        </div>
-      </div>
-
-      <div className="relative mt-4 flex flex-wrap items-center justify-between gap-3">
-        <button
-          type="button"
-          onClick={(e) => {
-            e.stopPropagation();
-            openMaps();
-          }}
-          disabled={!mapsUrl}
-          className="inline-flex items-center gap-2 rounded-full px-4 py-2.5 text-[13px] font-bold transition-colors disabled:opacity-50"
-          style={{
-            background: 'rgba(255,255,255,0.72)',
-            color: '#5A5070',
-            backdropFilter: 'blur(10px)',
-            WebkitBackdropFilter: 'blur(10px)',
-            fontFamily: 'Quicksand, ui-sans-serif, system-ui, sans-serif',
-          }}
-        >
-          Open in Maps
-          <span aria-hidden="true">↗</span>
-        </button>
-        {merchantId && (
-          <Link
-            {...merchantLink(merchantId)}
-            data-testid="receipt-view-all-at-location"
-            className="inline-flex items-center gap-1 text-[13px] font-semibold"
-            style={{
-              color: '#7A6F92',
-              fontFamily: 'Quicksand, ui-sans-serif, system-ui, sans-serif',
-            }}
-          >
-            View all visits
-            <span aria-hidden="true">→</span>
-          </Link>
-        )}
-      </div>
-    </div>
-  );
-}
-
-function FieldsGrid({
-  payment,
-  tax,
-  tip,
-  isProcessing,
-}: {
-  payment: string | null;
-  tax: number | undefined;
-  tip: number | undefined;
-  isProcessing: boolean;
-}) {
-  if (isProcessing) return null;
-  const cells: React.ReactNode[] = [];
-  if (tax != null && tax > 0) {
-    cells.push(<SmallFieldCard key="tax" label="Tax" value={`$${tax.toFixed(2)}`} numeric />);
-  }
-  if (tip != null && tip > 0) {
-    cells.push(<SmallFieldCard key="tip" label="Tip" value={`$${tip.toFixed(2)}`} numeric />);
-  }
-  if (payment) {
-    cells.push(
-      <SmallFieldCard
-        key="payment"
-        label="Payment"
-        value={payment.replace(/_/g, ' ').replace(/\b\w/g, (c) => c.toUpperCase())}
-      />,
-    );
-  }
-  if (cells.length === 0) return null;
-  return <div className="grid grid-cols-2 gap-3">{cells}</div>;
-}
-
-function SmallFieldCard({
-  label,
-  value,
-  numeric = false,
-}: {
-  label: string;
-  value: string;
-  numeric?: boolean;
-}) {
-  return (
-    <div className="rounded-[14px] border border-[var(--color-rule)] bg-[var(--color-surface)] px-4 py-3">
-      <p className="text-[11px] font-medium tracking-[0.14em] uppercase text-[var(--color-ink-muted)]">
-        {label}
-      </p>
-      <p
-        className={cn(
-          'mt-1 text-[15px] font-medium',
-          numeric && 'font-display italic font-medium text-lg tnum',
-        )}
-      >
-        {value}
-      </p>
-    </div>
-  );
-}
-
-// Subtle background per item class — keeps the badge informative
-// without screaming. Matches the paper palette.
-const ITEM_CLASS_STYLE: Record<BackendTransactionItem['item_class'], string> = {
-  durable: 'bg-[var(--color-paper-deep)] text-[var(--color-ink)]',
-  consumable: 'bg-amber-50 text-amber-900',
-  food_drink: 'bg-rose-50 text-rose-900',
-  service: 'bg-sky-50 text-sky-900',
-  other: 'bg-stone-100 text-stone-700',
-};
-
-const ITEM_CLASS_LABEL: Record<BackendTransactionItem['item_class'], string> = {
-  durable: 'durable',
-  consumable: 'consumable',
-  food_drink: 'food',
-  service: 'service',
-  other: 'other',
-};
-
-function formatMinor(minor: number | null, currency: string): string {
-  if (minor == null) return '—';
-  const sym = currency === 'USD' ? '$' : '';
-  return `${sym}${(minor / 100).toFixed(2)}`;
-}
-
-function LineItemsCard({
-  items,
-  currency,
-}: {
-  items: BackendTransactionItem[];
-  currency: string;
-}) {
-  return (
-    <div className="rounded-[18px] border border-[var(--color-rule)] bg-[var(--color-surface)] overflow-hidden">
-      <div className="px-5 py-4 border-b border-[var(--color-rule)] flex items-baseline justify-between">
-        <h3 className="font-display italic font-medium text-lg leading-none">
-          Items <span className="text-[var(--color-ink-muted)]">({items.length})</span>
-        </h3>
-        <span className="text-[10px] tracking-[0.16em] uppercase text-[var(--color-ink-muted)]">
-          from transaction_items
-        </span>
-      </div>
-      <ul className="divide-y divide-[var(--color-rule-soft)]">
-        {items.map((item) => {
-          const name = item.normalized_name?.trim() || item.raw_name;
-          return (
-            <li
-              key={item.line_no}
-              className="grid grid-cols-[1fr_auto] items-start gap-4 px-5 py-3"
-            >
-              <div className="min-w-0">
-                <div className="flex items-center gap-2">
-                  <span className="text-sm font-medium truncate">{name}</span>
-                  <span
-                    className={cn(
-                      'inline-flex items-center px-1.5 py-0.5 rounded-full text-[10px] font-medium leading-none',
-                      ITEM_CLASS_STYLE[item.item_class],
-                    )}
-                  >
-                    {ITEM_CLASS_LABEL[item.item_class]}
-                  </span>
-                  {item.line_type !== 'product' && (
-                    <span className="text-[10px] text-[var(--color-ink-muted)] uppercase tracking-wider">
-                      {item.line_type}
-                    </span>
-                  )}
-                </div>
-                {item.unit_price_minor != null && item.quantity !== 1 && (
-                  <div className="mt-0.5 text-[11px] text-[var(--color-ink-muted)] tnum">
-                    {item.quantity} × {formatMinor(item.unit_price_minor, item.currency || currency)}
-                  </div>
-                )}
-              </div>
-              <span className="font-display italic font-medium text-base tnum">
-                {formatMinor(item.line_total_minor, item.currency || currency)}
-              </span>
-            </li>
-          );
-        })}
-      </ul>
-    </div>
-  );
-}
-
-function ExtractionDetailsCollapsible({
-  rawText,
-  confidence,
-  warnings,
-  defaultOpen,
-}: {
-  rawText: string | undefined;
-  confidence: number | undefined;
-  warnings: string[] | undefined;
-  defaultOpen: boolean;
-}) {
-  const [open, setOpen] = useState(defaultOpen);
-  const isLow = confidence != null && confidence < 0.6;
-  return (
-    <div className="rounded-[18px] border border-[var(--color-rule)] bg-[var(--color-surface)] overflow-hidden">
-      <button
-        type="button"
-        onClick={() => setOpen((s) => !s)}
-        className="w-full px-5 py-4 flex items-center justify-between text-left hover:bg-[var(--color-paper-deep)]/30 transition-colors"
-      >
-        <span className="font-display italic font-medium text-lg leading-none">
-          Extraction details
-        </span>
-        <span className="flex items-center gap-2">
-          {confidence != null && (
-            <span
-              className={cn(
-                'text-[11px] tracking-[0.12em] uppercase',
-                isLow ? 'text-[var(--color-stamp)]' : 'text-[var(--color-ink-muted)]',
-              )}
-            >
-              {(confidence * 100).toFixed(0)}% confidence
-            </span>
-          )}
-          {open ? <ChevronDown size={18} className="text-[var(--color-ink-muted)]" /> : <ChevronRight size={18} className="text-[var(--color-ink-muted)]" />}
-        </span>
-      </button>
-      {open && (
-        <div className="px-5 pb-5 space-y-4">
-          {warnings && warnings.length > 0 && (
-            <div className="flex flex-wrap gap-1">
-              {warnings.map((w, i) => (
-                <span
-                  key={i}
-                  className="px-2 py-0.5 rounded-full bg-[var(--color-stamp)]/10 text-[var(--color-stamp)] text-[10px] tracking-[0.08em] uppercase"
-                >
-                  {w}
-                </span>
-              ))}
-            </div>
-          )}
-          {rawText && (
-            <pre className="text-xs text-[var(--color-ink-muted)] whitespace-pre-wrap font-mono">
-              {rawText}
-            </pre>
-          )}
-        </div>
-      )}
-    </div>
-  );
-}
-
-function ProcessingNote() {
-  return (
-    <div
-      className={cn(
-        'rounded-[16px] px-4 py-4 flex items-start gap-3',
-        'border border-[var(--color-rule)] bg-[var(--color-butter)]/40',
-      )}
-    >
-      <span
-        aria-hidden="true"
-        className="mt-1 inline-block h-2.5 w-2.5 rounded-full bg-[var(--color-terracotta)] animate-pulse"
-      />
-      <div>
-        <p className="font-display italic font-medium">Still reading your receipt</p>
-        <p className="mt-1 text-sm text-[var(--color-ink-muted)]">
-          Claude is extracting fields — this page will refresh on its own.
-        </p>
-      </div>
-    </div>
-  );
-}
-
-function Banner({
-  tone,
-  children,
-}: {
-  tone: 'error';
-  children: React.ReactNode;
-}) {
-  return (
-    <div
-      className={cn(
-        'rounded-[16px] px-4 py-3 text-sm',
-        tone === 'error'
-          ? 'border border-[var(--color-stamp)]/30 bg-[var(--color-stamp)]/5 text-[var(--color-stamp)]'
-          : '',
-      )}
-    >
-      {children}
-    </div>
-  );
-}
-
-/**
- * Feedback banner for re-extract. Mirrors the RefreshBanner pattern
- * in MerchantDetail — same aesthetic, separate copy to avoid pulling
- * a tiny presentational helper across the file boundary. If a third
- * surface needs the same widget, factor it out then.
- */
-function ReExtractBanner({
-  tone,
-  children,
-  onDismiss,
-}: {
-  tone: 'success' | 'error';
-  children: React.ReactNode;
-  onDismiss: () => void;
-}) {
-  return (
-    <div
-      className={cn(
-        'flex items-start justify-between gap-3 rounded-[14px] px-4 py-3 text-sm',
-        tone === 'success' &&
-          'border border-[var(--color-terracotta)]/30 bg-[var(--color-terracotta)]/8 text-[var(--color-ink)]',
-        tone === 'error' &&
-          'border border-[var(--color-stamp)]/40 bg-[var(--color-stamp)]/5 text-[var(--color-stamp)]',
-      )}
-    >
-      <p className="font-hand text-base leading-snug">{children}</p>
-      <button
-        type="button"
-        onClick={onDismiss}
-        className="text-[11px] uppercase tracking-[0.16em] opacity-60 hover:opacity-100 transition-opacity shrink-0 mt-0.5"
-        aria-label="Dismiss"
-      >
-        dismiss
-      </button>
-    </div>
-  );
-}
-
-/* ── Helpers ──────────────────────────────────────────────────── */
-
-function formatDateLong(isoDate: string): string {
-  const [y, m, d] = isoDate.split('-').map(Number);
-  if (!Number.isFinite(y) || !Number.isFinite(m) || !Number.isFinite(d)) return isoDate;
-  const dt = new Date(y, m - 1, d);
-  return dt.toLocaleString('en-US', { month: 'short', day: 'numeric', year: 'numeric' });
-}
